some major changes(incomplete)

+++ src/entity/Entity.java
@@ -34,7 +34,7 @@ public abstract class Entity {
         return posX;
     }
 
-    public void setPosX(int posX) {
+    public void setPosX(double posX) {
         this.posX = posX;
     }
 
@@ -42,7 +42,7 @@ public abstract class Entity {
         return posY;
     }
 
-    public void setPosY(int posY) {
+    public void setPosY(double posY) {
         this.posY = posY;
     }
 }
